Explicit boxing
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/jcs/trunk@1584226 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/commons/jcs/auxiliary/disk/block/BlockDisk.java
@@ -148,9 +148,9 @@ public class BlockDisk
             Integer emptyBlock = emptyBlocks.takeFirst();
             if (emptyBlock == null)
             {
-                emptyBlock = numberOfBlocks.getAndIncrement();
+                emptyBlock = Integer.valueOf(numberOfBlocks.getAndIncrement());
             }
-            blocks[i] = emptyBlock;
+            blocks[i] = emptyBlock.intValue();
         }
 
         return blocks;
